Fix OBM tests - Delete redundant test_node_id_obm_create() - Move OBM identify test into OBMTests() and set check-obm dependency
