+++ test/tests/obm_tests.py
@@ -1,4 +1,6 @@
 from config.settings import *
+from on_http import NodesApi as Nodes
+from on_http import rest
 from modules.obm import obmSettings
 from modules.logger import Log
 from datetime import datetime
@@ -9,6 +11,7 @@ from proboscis.asserts import assert_true
 from proboscis.asserts import assert_not_equal
 from proboscis import SkipTest
 from proboscis import test
+from json import loads, dumps
 
 LOG = Log(__name__)
 
@@ -16,8 +19,8 @@ LOG = Log(__name__)
 class OBMTests(object):
 
     def __init__(self):
-        pass
-    
+        self.__client = config.api_client 
+
     @test(groups=['obm.tests', 'set-ipmi-obm'], depends_on_groups=['nodes.tests'])
     def setup_ipmi_obm(self):
         """ Setup IPMI OBM settings with PATCH:/nodes """
@@ -40,4 +43,20 @@ class OBMTests(object):
         assert_not_equal(len(obmSettings().check_nodes(service_type='snmp-obm-service')), 0,
                 message='there are missing SNMP OBM settings!')
    
+    @test(groups=['obm.tests', 'create-node-id-obm-identify'], depends_on_groups=['check-obm'])
+    def test_node_id_obm_identify_create(self):
+        """ Testing POST:/nodes/:id/obm/identify """
+        Nodes().api1_1_nodes_get()
+        nodes = loads(self.__client.last_response.data)
+        codes = []
+        data = {"value": "true"}
+        for n in nodes:
+            if n.get('type') == 'compute':
+                uuid = n.get('id')
+                Nodes().api1_1_nodes_identifier_obm_identify_post(uuid, data)
+                rsp = self.__client.last_response
+                codes.append(rsp)
+        for c in codes:
+            assert_equal(200, c.status, message=c.reason)
+        assert_raises(rest.ApiException, Nodes().api1_1_nodes_identifier_obm_identify_post, 'fooey', data)
